backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Lucas Franco",email = "127855437+lucasfc-dev@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "mutagen (>=1.47.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
